changes to make compatible with jsdom 2.0.0-4.0.1
When trying to use the code with jsdom v3 or v4, I found that I needed to make some small changes to the setup code.

I tested the code against 2.0.0 and it works as well, so it would appear to be safe to switch to this for better compatibility with newer versions.

global.navigator was needed when I started testing with events.